fixed a compiler warning
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@2004 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ io/src/io.cpp
@@ -89,7 +89,7 @@ pcl::concatenateFields (const sensor_msgs::PointCloud2 &cloud1,
       cloud1_unique_fields.push_back(&(cloud1.fields[i]));
 
       int size = 0;
-      int next_valid_field = i + 1;
+      size_t next_valid_field = i + 1;
 
       while(next_valid_field < cloud1.fields.size())
       {
